100% coverage

+++ kata-solutions/pacman-kata/pacman-python/src/adapters/pygame_clock.py
@@ -6,5 +6,5 @@ class PyGameClock(Clock):
     def __init__(self):
         self._clock = pygame.time.Clock()
 
-    def tick(self, rate: int) -> int:
+    def tick(self, rate: float) -> int:
         return self._clock.tick()
